New Commit 17/07/2025: Playlist Update 6
- Added new Liked Songs playlist that is made for all users. This playlist can't be deleted but editing

- Ensured that no other playlist can be renamed to the same as Liked Songs

- Added heart button for liking songs. When a song is liked, it will be added to

BUG(s):
 - When inside the liked songs playlist, if I unlike a song, it doesn't automatically refresh the playlist showing the updated playlist with the unliked song removed.

NEXT UPDATE:
 - Look to add more songs potentially using Spotify's api
 - Setup backend properties for listening to music
 - Add in Settings for users to edit their name

+++ app/backend/routes/playlists.py
@@ -26,7 +26,12 @@ def create_playlist(
         db: Session = Depends(get_db),
         current_user = Depends(get_current_user)
 ):
-    pl = Playlist(name=playlist_in.name, user_id=current_user.id)
+
+    # Prevent the creation of another Like Songs playlist
+    if playlist_in.name.lower() == "Liked Songs":
+        raise HTTPException(status_code=400, detail="Cannot create another playlist called 'Liked Songs' ")
+
+    pl = Playlist(name=playlist_in.name, user_id=current_user.id, is_liked_songs=False)
     db.add(pl)
     db.commit()
     db.refresh(pl)
@@ -44,6 +49,15 @@ def rename_playlist(playlist_id: int, update: PlaylistUpdate, db: Session = Depe
     pl = db.get(Playlist, playlist_id)
     if not pl or pl.user_id != current_user.id:
         raise HTTPException(status_code=404, detail="Playlist not found")
+
+    # Prevent renaming Liked Songs playlist
+    if pl.is_liked_songs:
+        raise HTTPException(status_code=400, detail="Cannot rename to 'Liked Songs'")
+
+    # Prevent renaming to "Liked Songs"
+    if update.name.lower() == "Liked Songs":
+        raise HTTPException(status_code=400, detail="Cannot rename to 'Liked Songs'")
+
     pl.name = update.name
     db.add(pl); db.commit(); db.refresh(pl)
     return pl
@@ -53,6 +67,11 @@ def delete_playlist(playlist_id: int, db: Session = Depends(get_db), current_use
     pl = db.get(Playlist, playlist_id)
     if not pl or pl.user_id != current_user.id:
         raise HTTPException(status_code=404, detail="Playlist not found")
+
+    # Prevent the deletion of "Liked Songs"
+    if pl.is_liked_songs:
+        raise HTTPException(status_code=400, detail="Cannot delete 'Liked Songs'")
+
     db.delete(pl); db.commit();
     return
 
@@ -96,3 +115,23 @@ def remove_track_from_playlist(
     pl.songs.remove(song)
     db.add(pl); db.commit(); db.refresh(pl)
     return {"detail": "Track removed"}
+
+# Get liked songs of the user
+@router.get("/special/liked-songs", response_model= PlaylistDetail)
+def get_liked_songs_playlist(current_user: User = Depends(get_current_user), db: Session = Depends(get_db)):
+    stmt = select(Playlist).where(
+        Playlist.user_id == current_user.id,
+        Playlist.is_liked_songs == True,
+    )
+    liked_playlist = db.exec(stmt).first()
+
+    if not liked_playlist:
+        # Create a liked playlist if it doesn't exist
+        liked_playlist = Playlist(
+            name="Liked Songs",
+            user_id=current_user.id,
+            is_liked_songs=True,
+        )
+        db.add(liked_playlist); db.commit(); db.refresh(liked_playlist)
+
+    return liked_playlist
